+++ src/redux/selectors/work.js
@@ -1,13 +1,20 @@
 import { createOrmSelector, ormSessionSelector } from './orm';
 
 export const runningSessionIdSelector = state => state.work.runningSessionId;
+export const lastRunningSessionIdSelector = state => state.work.lastRunningSessionId;
 const selectedSubjectIdSelector = state => state.work.selectedSubjectId;
 
 export const runningSessionSelector = createOrmSelector(
   ormSessionSelector,
   runningSessionIdSelector,
-  (ormSession, runningSessionId) => runningSessionId !== -1
-    && ormSession.WorkSession.withId(runningSessionId),
+  (ormSession, id) => id !== -1 && ormSession.WorkSession.withId(id),
+);
+
+export const lastRunningSessionSelector = createOrmSelector(
+  ormSessionSelector,
+  lastRunningSessionIdSelector,
+  (ormSession, id) => id !== -1 && ormSession.WorkSession.idExists(id)
+    && ormSession.WorkSession.withId(id),
 );
 
 export const selectedSubjectSelector = createOrmSelector(